fix(project): move generate_project_name to project model
